+++ lib/src/responses/search_response.dart
@@ -3,10 +3,9 @@ import 'package:meta/meta.dart';
 
 import '../utilities/helpers.dart';
 import '../utilities/self_representive_base.dart';
-import 'properties/links.dart';
 
 @immutable
-class Search implements ISelfRespresentive {
+final class Search implements ISelfRespresentive {
   const Search({
     required this.id,
     required this.type,
@@ -14,20 +13,15 @@ class Search implements ISelfRespresentive {
     required this.url,
     required this.self,
     this.title,
-    this.links,
   });
 
   factory Search.fromJson(Map<String, dynamic> json) {
     return Search(
-      id: castOrElse(json['id']),
+      id: castOrElse(json['id'])!,
       title: castOrElse(json['title']),
       type: castOrElse(json['type']),
       subType: castOrElse(json['subtype']),
       url: castOrElse(json['url']),
-      links: castOrElse(
-        json['_links'],
-        transformer: (value) => Links.fromJson(value as Map<String, dynamic>),
-      ),
       self: json,
     );
   }
@@ -36,7 +30,6 @@ class Search implements ISelfRespresentive {
   final String? title;
   final String type;
   final String subType;
-  final Links? links;
   final String url;
 
   @override
@@ -49,7 +42,6 @@ class Search implements ISelfRespresentive {
       'type': type,
       'subtype': subType,
       'url': url,
-      '_links': links?.toJson(),
     };
   }
 
@@ -65,7 +57,6 @@ class Search implements ISelfRespresentive {
         other.title == title &&
         other.type == type &&
         other.subType == subType &&
-        other.links == links &&
         other.url == url &&
         mapEquals(other.self, self);
   }
@@ -76,13 +67,12 @@ class Search implements ISelfRespresentive {
         title.hashCode ^
         type.hashCode ^
         subType.hashCode ^
-        links.hashCode ^
         url.hashCode ^
         self.hashCode;
   }
 
   @override
   String toString() {
-    return 'Search(id: $id, title: $title, type: $type, subType: $subType, links: $links, url: $url, self: $self)';
+    return 'Search(id: $id, title: $title, type: $type, subType: $subType, url: $url, self: $self)';
   }
 }